+++ src/main/java/http/helpers/DataHelper.java
@@ -3,7 +3,7 @@ package http.helpers;
 import java.util.HashMap;
 
 /**
- * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.http.helpers.DataHelper}
+ * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.helpers.DataHelper}
  */
 @Deprecated()
 public class DataHelper {
