+++ src/main/java/http/helpers/EnvironmentHelper.java
@@ -3,13 +3,13 @@ package http.helpers;
 import java.util.Map;
 
 /**
- * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.http.helpers.EnvironmentHelper}
+ * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.helpers.EnvironmentHelper}
  */
 @Deprecated()
 public class EnvironmentHelper {
 
     public static void setEnv(Map<String, String> newenv) {
-        net.itarray.automotion.tools.http.helpers.EnvironmentHelper.setEnv(newenv);
+        net.itarray.automotion.tools.helpers.EnvironmentHelper.setEnv(newenv);
     }
 
 }
