+++ src/main/java/net/itarray/automotion/tools/helpers/EnvironmentHelper.java
@@ -1,4 +1,4 @@
-package net.itarray.automotion.tools.http.helpers;
+package net.itarray.automotion.tools.helpers;
 
 import net.itarray.automotion.tools.environment.EnvironmentConstants;
