+++ src/main/java/net/itarray/automotion/tools/helpers/MailService.java
@@ -1,4 +1,4 @@
-package net.itarray.automotion.tools.http.helpers;
+package net.itarray.automotion.tools.helpers;
 
 import javax.mail.*;
 import java.util.Properties;
