+++ src/main/java/net/itarray/automotion/tools/helpers/TextFinder.java
@@ -1,4 +1,4 @@
-package net.itarray.automotion.tools.http.helpers;
+package net.itarray.automotion.tools.helpers;
 
 import java.util.Arrays;
 import java.util.Map;
